Display ordered lists and tables properly in the read-only rich-text editor (#14)
The `<AuContent>` component doesn't support `<ol>` or tables yet. A fix is on its way, but we're still on Appuniversum v2 so we would need to update to get it.

As a temporary workaround I included the same styles in the shame.scss file. We can remove them again when we update Appuniversum.