Fix a "three-way-compare" link styling issue (#15)
In a previous commit we replaced the `AuLink` with a `AuButton` since that's semantically correct element to use for this. However, this causes styling conflicts with the CSS hacks the project uses to style the delete buttons in the form.

So we revert back to a link, but with a regular `<a>` element and with a comment explaining the situation.